Added ability to use gnubg hint feature as opposed to the external interface.
As it turns out, the external is broken. Now that MatchDog can export
to sgf format and analyse matches, it's crystal clear that it sometimes
gets incorrect responses for both rollout and cube decision from external.
In some instances swapping direction and color in such a case results
in a correct response but it's  unpredictable and inconsistent.
In some cases both the original and the swapped board gives incorrect responses.

So instead of cleaning up the terrible mess in FibsRunner and elsewhere
I just made it worse but now it plays using hints by default. Configurable
in player preferences. Resumed matches are not yet supported for hints so
we fall back to external.